requirements.txt
pyyaml
torch
pymilvus>=2.4.0
sentence-transformers
numpy
tqdm
